requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
twilio==6.45.0
Werkzeug==1.0.1
WTForms==2.3.3
torch==1.0.0a0+ff608a9
torchvision==0.2.2
imutils
pytesseract